CsvTableView: consider separators as special chars even with quotes
Formerly record and field separators where not considered as special
chars when a quote character was setted.
In other words, with default CSV separators, \n and , were written
within fields as soon as quoting was activating, leading to CSV files
harder to read for most parsers, and notably not correctly read by
Microsoft Excel.