thread: simplified circular buffer and moved to thread directory
Having a "lock-less" circular buffer using QSemaphore when only one
thread writes or reads from the buffer is uselss since QSemaphore
actually relies on QMutex and therferore is not lock-less.

Moved CircularBuffer and AtomicValue to thread directory, which is
consistent with Qt choices (same as QSemaphore or QWaitCondition),
moreover CircularBuffer is an inter-thread communication mean, not
a container.